refactor(racing-car): racing-car를 domain으로 이동, 책임과 역할에 따른 수정
